+++ json-schema/python/frinx_api/json_schema/device_geo_location.py
@@ -2,11 +2,20 @@
 
 from __future__ import annotations
 
+from enum import Enum
 from typing import Optional
 
 from pydantic import BaseModel
 from pydantic import ConfigDict
 from pydantic import Field
+
+
+class Type(str, Enum):
+    """
+    Type of geometry.
+    """
+
+    Point = 'Point'
 
 
 class DeviceGeoLocation(BaseModel):
@@ -17,7 +26,7 @@ class DeviceGeoLocation(BaseModel):
     model_config = ConfigDict(
         populate_by_name=True,
     )
-    type: Optional[str] = 'Point'
+    type: Type
     """
     Type of geometry.
     """